Update Artiodactyla clustering analysis.ipynb
Finished functions to plot and save results from MANOVAs. The purpose of these function is to see how well different catagories (sex, Genus, family, species) can be distinguished. In addition, within each catagory, each group was compared to all th other groups combined, as a way to see which groups had higher variance, or were more different from other groups, ect.
A future step with the MANOVA could be to look at a subset of the data, so that we are restricting comparisions between sexes to data with a single family or genus, rather than within artiodactyla as a whole.
Also, finishing the machine learning  predictions is a high priority.
Longer term goals could include adding the full dataset. Unfortunatly this data is the form of images. Not sure if possible to automatically take these measurement from the images of if it will be a manual task.